requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
supabase>=2.0.0
python-docx>=1.1.2
